[Fix issues in lib]
