--- Example/example/Utility/PlistKeys.swift
@@ -1,14 +1,0 @@
-//
-//  PlistUtility.swift
-//  Example
-//
-//  Created by Talha Rasool on 20/04/2025.
-//
-
-import Foundation
-
-
-
-enum PlistKey: String {
-    case PUBLISHED_API_KEY = "PUBLISHED_API_KEY"
-}
